Fix and adjust Omni-Tool mining speed
The Omni-Tool has a rather subtle bug that shows up when you mine
blocks with a high hardness.

Since this weapon doesn't use the default block breaking mechanics
they had to be reimplemented. This was done poorly.

The author tried to implement scaling based on the damage modifier but
this was done in a way that caused the multiplier to be applied to the
current damage rather than the damage that gets applied that tick.

This causes multiplicative scaling on the current damage making it go
up much quicker than it should.

This is different from the default methods causing a de-sync between
the client and the server by making the client have it's mining go
much faster.
I correct this by:

1. Doing the damage scaling in the new getDigSpeed() method.
2. Correcting the partial block breaking to what the default methods use.

This exposed the rather shameful default mining speed. I decided to
change it to 9 after some in game testing. This is 3x the speed of an
un-enchanted diamond pickaxe and 1.5x of the tritanium pickaxe.

This speed is rather quick whilst not making it trivial to get over the
"instant mining" threshold.

These changes aren't particularly clean in my opinion but they work
and don't seem to have any nasty side effects.