Disable default matter registration entirely
They just aren't used in the pack and as such disabling them here is much
easier than blacklisting every default item.

Really need to make this an actual config option but the config handler
scares me so that is for later.

Also switched some random loop with an "enhanced for loop".
I have no idea what that is but the IDE seems to like it so, sure I guess.

Fixed some warnings in "MatterRegistryCommands.java". Didn't seem to help
with the entire config wiping itself whenever you try to add a second
blacklist entry via the command. Really need to get to that at some point.