Stops the blacklist command from wiping the config when entries exist
This command needs to expand the array that holds the blacklist strings to be written to the config.
It does this by creating a new array with 1 extra length but forgets to copy the contents of the old array.
This creates a bunch of slots with a null value which causes things to break somewhere in the saving process making the config wipe itself.
I fixed this by copying the contents of the old array to the new one preventing the null entries and it gets written like it should.